move get_dynamic_ins to discovery.py
